refactor: announcer 훅 context api로 전환
