+++ src/index.tsx
@@ -1,10 +1,10 @@
 import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
 import App from './App';
+import { AnnouncerProvider } from './contexts/AnnouncerContext';
 
 ReactDOM.render(
-  <BrowserRouter>
+  <AnnouncerProvider>
     <App />
-  </BrowserRouter>,
+  </AnnouncerProvider>,
   document.getElementById('root')
 );
